refactor(authenticationContext): uses context on navbar and moves logic to UserBubble. It also uses a skeleton
